requirements.txt
python-telegram-bot==21.0.1
spotdl==4.2.4
python-dotenv==1.0.1